refactor responses funcs to /grpc/responses
