+++ sgl-router/src/routers/grpc/router.rs
@@ -9,29 +9,19 @@ use axum::{
     http::{HeaderMap, StatusCode},
     response::{IntoResponse, Response},
 };
-use serde_json::json;
-use tokio::{sync::RwLock, task::JoinHandle};
-use tracing::{debug, warn};
+use tokio::sync::RwLock;
+use tracing::debug;
 
-/// Information stored for background tasks to enable end-to-end cancellation
-///
-/// This struct enables cancelling both the Rust task AND the Python scheduler processing.
-/// The client field is lazily initialized during pipeline execution.
-pub struct BackgroundTaskInfo {
-    /// Tokio task handle for aborting the Rust task
-    pub handle: JoinHandle<()>,
-    /// gRPC request_id sent to Python scheduler (chatcmpl-* prefix)
-    pub grpc_request_id: String,
-    /// gRPC client for sending abort requests to Python (set after client acquisition)
-    pub client: Arc<RwLock<Option<crate::grpc_client::SglangSchedulerClient>>>,
-}
-
-use super::{context::SharedComponents, pipeline::RequestPipeline, responses};
+use super::{
+    context::SharedComponents,
+    pipeline::RequestPipeline,
+    responses::{self, BackgroundTaskInfo},
+};
 use crate::{
     config::types::RetryConfig,
     core::WorkerRegistry,
     data_connector::{
-        ResponseId, SharedConversationItemStorage, SharedConversationStorage, SharedResponseStorage,
+        SharedConversationItemStorage, SharedConversationStorage, SharedResponseStorage,
     },
     policies::PolicyRegistry,
     protocols::{
@@ -306,187 +296,11 @@ impl RouterTrait for GrpcRouter {
         response_id: &str,
         _params: &ResponsesGetParams,
     ) -> Response {
-        let resp_id = ResponseId::from(response_id);
-
-        // Retrieve response from storage
-        match self.response_storage.get_response(&resp_id).await {
-            Ok(Some(stored_response)) => {
-                axum::Json(stored_response.raw_response).into_response()
-            }
-            Ok(None) => {
-                (
-                    StatusCode::NOT_FOUND,
-                    axum::Json(json!({
-                        "error": {
-                            "message": format!("Response with id '{}' not found", response_id),
-                            "type": "not_found_error",
-                            "code": "response_not_found"
-                        }
-                    })),
-                )
-                    .into_response()
-            }
-            Err(e) => {
-                (
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    axum::Json(json!({
-                        "error": {
-                            "message": format!("Failed to retrieve response: {}", e),
-                            "type": "internal_error"
-                        }
-                    })),
-                )
-                    .into_response()
-            }
-        }
+        responses::get_response_impl(&self.response_storage, response_id).await
     }
 
     async fn cancel_response(&self, _headers: Option<&HeaderMap>, response_id: &str) -> Response {
-        let resp_id = ResponseId::from(response_id);
-
-        // Retrieve response from storage to check if it exists and get current status
-        match self.response_storage.get_response(&resp_id).await {
-            Ok(Some(stored_response)) => {
-                // Check current status - only queued or in_progress responses can be cancelled
-                let current_status = stored_response
-                    .raw_response
-                    .get("status")
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("unknown");
-
-                match current_status {
-                    "queued" | "in_progress" => {
-                        // Attempt to abort the background task
-                        let mut tasks = self.background_tasks.write().await;
-                        if let Some(task_info) = tasks.remove(response_id) {
-                            // Abort the Rust task immediately
-                            task_info.handle.abort();
-
-                            // Abort the Python/scheduler request via gRPC (if client is available)
-                            let client_opt = task_info.client.read().await;
-                            if let Some(ref client) = *client_opt {
-                                if let Err(e) = client
-                                    .abort_request(
-                                        task_info.grpc_request_id.clone(),
-                                        "User cancelled via API".to_string(),
-                                    )
-                                    .await
-                                {
-                                    warn!(
-                                        "Failed to abort Python request {}: {}",
-                                        task_info.grpc_request_id, e
-                                    );
-                                } else {
-                                    debug!(
-                                        "Successfully aborted Python request: {}",
-                                        task_info.grpc_request_id
-                                    );
-                                }
-                            } else {
-                                debug!("Client not yet available for abort, request may not have started yet");
-                            }
-
-                            // Task was found and aborted
-                            (
-                                StatusCode::OK,
-                                axum::Json(json!({
-                                    "id": response_id,
-                                    "status": "cancelled",
-                                    "message": "Background task has been cancelled"
-                                })),
-                            )
-                                .into_response()
-                        } else {
-                            // Task handle not found (may have already completed)
-                            (
-                                StatusCode::OK,
-                                axum::Json(json!({
-                                    "id": response_id,
-                                    "status": "completed_or_not_found",
-                                    "message": "Task may have already completed before cancellation"
-                                })),
-                            )
-                                .into_response()
-                        }
-                    }
-                    "completed" => {
-                        (
-                            StatusCode::BAD_REQUEST,
-                            axum::Json(json!({
-                                "error": {
-                                    "message": "Cannot cancel completed response",
-                                    "type": "invalid_request_error",
-                                    "code": "response_already_completed"
-                                }
-                            })),
-                        )
-                            .into_response()
-                    }
-                    "failed" => {
-                        (
-                            StatusCode::BAD_REQUEST,
-                            axum::Json(json!({
-                                "error": {
-                                    "message": "Cannot cancel failed response",
-                                    "type": "invalid_request_error",
-                                    "code": "response_already_failed"
-                                }
-                            })),
-                        )
-                            .into_response()
-                    }
-                    "cancelled" => {
-                        (
-                            StatusCode::OK,
-                            axum::Json(json!({
-                                "id": response_id,
-                                "status": "cancelled",
-                                "message": "Response was already cancelled"
-                            })),
-                        )
-                            .into_response()
-                    }
-                    _ => {
-                        // Unknown status
-                        (
-                            StatusCode::INTERNAL_SERVER_ERROR,
-                            axum::Json(json!({
-                                "error": {
-                                    "message": format!("Unknown response status: {}", current_status),
-                                    "type": "internal_error"
-                                }
-                            })),
-                        )
-                            .into_response()
-                    }
-                }
-            }
-            Ok(None) => {
-                (
-                    StatusCode::NOT_FOUND,
-                    axum::Json(json!({
-                        "error": {
-                            "message": format!("Response with id '{}' not found", response_id),
-                            "type": "not_found_error",
-                            "code": "response_not_found"
-                        }
-                    })),
-                )
-                    .into_response()
-            }
-            Err(e) => {
-                (
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    axum::Json(json!({
-                        "error": {
-                            "message": format!("Failed to retrieve response: {}", e),
-                            "type": "internal_error"
-                        }
-                    })),
-                )
-                    .into_response()
-            }
-        }
+        responses::cancel_response_impl(&self.response_storage, &self.background_tasks, response_id).await
     }
 
     async fn route_classify(
